+++ lib/src/check_command.dart
@@ -15,12 +15,11 @@ import 'package:analyzer/source/analysis_options_provider.dart';
 import 'package:analyzer/source/sdk_ext.dart';
 import 'package:analyzer/src/dart/sdk/sdk.dart';
 import 'package:analyzer/src/generated/engine.dart';
-import 'package:analyzer/src/generated/error.dart';
+import 'package:analyzer/src/generated/error.dart'; // ignore: deprecated_member_use
 import 'package:analyzer/src/generated/java_io.dart';
 import 'package:analyzer/src/generated/sdk.dart';
 import 'package:analyzer/src/generated/source.dart';
 import 'package:analyzer/src/generated/source_io.dart';
-import 'package:analyzer/src/task/options.dart';
 import 'package:package_config/discovery.dart' as pkgDiscovery;
 import 'package:package_config/packages.dart' show Packages;
 import 'package:path/path.dart' as p;
@@ -47,7 +46,6 @@ class CheckCommand extends Command {
     );
     AnalysisContext context = AnalysisEngine.instance.createAnalysisContext();
     AnalysisOptionsImpl options = new AnalysisOptionsImpl();
-    options.cacheSize = 512;
     AnalysisEngine.instance.processRequiredPlugins();
 
     List<UriResolver> resolvers = [
@@ -60,6 +58,7 @@ class CheckCommand extends Command {
       packages = _discoverPackagespec(project.dir);
       resolvers.add(new SdkExtUriResolver(_createPackageFilePackageMap(packages)));
     } else if (project.packageDir.existsSync()) {
+      // ignore: deprecated_member_use
       new PackageUriResolver([new JavaFile(project.packagePath)]);
       resolvers.add(new SdkExtUriResolver(_createPackagesFolderPackageMap(project)));
     }
@@ -145,7 +144,6 @@ class CheckCommand extends Command {
     return errors.isEmpty ? new Future.value() : new Future.error(new ExitCode(1));
   }
 
-
   Map<String, List<Folder>> _createPackageFilePackageMap(Packages packages) {
     Map<String, List<Folder>> m = {};
     Map<String, Uri> packageMap = packages.asMap();
@@ -199,9 +197,10 @@ class CheckCommand extends Command {
     if (options == null || options.isEmpty) return;
 
     // Handle options processors.
-    List processors = AnalysisEngine.instance.optionsPlugin.optionsProcessors;
-    processors.forEach((processor) => processor.optionsProcessed(context, options));
-    configureContextOptions(context, options);
+    // TODO(devoncarew): Cleanup as part of move to analyzer 0.30.0-alpha.1.
+    // List processors = AnalysisEngine.instance.optionsPlugin.optionsProcessors;
+    // processors.forEach((processor) => processor.optionsProcessed(context, options));
+    // configureContextOptions(context, options);
 
     // Handle strong mode.
     // analyzer:
@@ -231,7 +230,14 @@ class _Error implements Comparable {
   int get severity => error.errorCode.errorSeverity.ordinal;
   String get severityName => error.errorCode.errorSeverity.displayName;
   String get message => error.message;
-  String get description => '${message} at ${location}, line ${line}.';
+  String get messageSentenceFragment {
+    String m = message;
+    return m.endsWith('.') ? m.substring(0, m.length - 1) : m;
+  }
+  String get code => error.errorCode.name.toLowerCase();
+
+  String get description => '${messageSentenceFragment} at ${location}, line ${line} ($code).';
+
   int get line => lineInfo.getLocation(error.offset).lineNumber;
 
   String get location {